requirements.txt
flask 
torch 
torchvision
werkzeug
pillow